# swagger.json
{
  "x-generator": "NSwag v13.7.4.0 (NJsonSchema v10.1.26.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:44394",
  "schemes": [
    "https"
  ],
  "paths": {
    "/api/Person": {
      "get": {
        "tags": [
          "Person"
        ],
        "operationId": "Person_GetPersonAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Person"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Person"
        ],
        "operationId": "Person_PostPerson",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "person",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Person"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        }
      }
    },
    "/api/Person/{id}": {
      "get": {
        "tags": [
          "Person"
        ],
        "operationId": "Person_GetPerson",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Person"
        ],
        "operationId": "Person_PutPerson",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "person",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Person"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Person"
        ],
        "operationId": "Person_DeletePerson",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Person": {
      "type": "object",
      "required": [
        "id",
        "birthDate",
        "gender",
        "role"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "guid"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "birthDate": {
          "type": "string",
          "format": "date-time"
        },
        "login": {
          "type": "string"
        },
        "gender": {
          "$ref": "#/definitions/Gender"
        },
        "role": {
          "$ref": "#/definitions/PersonRole"
        }
      }
    },
    "Gender": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Male",
        "Female"
      ],
      "enum": [
        0,
        1
      ]
    },
    "PersonRole": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Patient",
        "Doctor",
        "Admin"
      ],
      "enum": [
        0,
        1,
        2
      ]
    }
  }
}